requirements.txt
streamlit
PyMuPDF
faiss-cpu
sentence-transformers
langchain
tqdm
ollama
pdfplumber
pandas
google-generativeai
python-dotenv
